+++ src/Auth/AuthMethod.php
@@ -8,4 +8,9 @@ enum AuthMethod: string
     case HYVOR = 'hyvor';
     case OIDC = 'oidc';
 
+    public function isOidc(): bool
+    {
+        return $this === self::OIDC;
+    }
+
 }