+++ tests/SymfonyTestCase.php
@@ -7,14 +7,19 @@ use Doctrine\DBAL\Connection;
 use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Symfony\Bundle\FrameworkBundle\Console\Application;
+use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
+use Zenstruck\Foundry\Test\Factories;
 
 // https://symfonycasts.com/screencast/symfony-bundle/integration-test
 class SymfonyTestCase extends TestCase
 {
+
+    use Factories;
 
     public SymfonyKernel $kernel;
     public Container $container;
@@ -51,6 +56,13 @@ class SymfonyTestCase extends TestCase
         return $this->container;
     }
 
+    protected function getCommandTester(string $command): CommandTester
+    {
+        $application = new Application($this->kernel);
+        $command = $application->find($command);
+        return new CommandTester($command);
+    }
+
     protected function createTables(): void
     {
         $doctrine = $this->container->get('doctrine');
@@ -59,6 +71,7 @@ class SymfonyTestCase extends TestCase
         assert($connection instanceof Connection);
 
         $connection->executeQuery('DROP TABLE IF EXISTS oidc_users;');
+        $connection->executeQuery('DROP TABLE IF EXISTS sudo_users;');
         $connection->executeQuery(
             <<<SQL
         CREATE TABLE oidc_users (
@@ -72,6 +85,15 @@ class SymfonyTestCase extends TestCase
           picture_url TEXT,
           website_url TEXT,
           UNIQUE (iss, sub)
+        );
+        SQL
+        );
+        $connection->executeQuery(
+            <<<SQL
+        CREATE TABLE sudo_users (
+            user_id INTEGER PRIMARY KEY,
+            created_at TEXT NOT NULL,
+            updated_at TEXT NOT NULL
         );
         SQL
         );